DJANGO/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
openai==1.30.1
openapi==1.1.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.18.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
